requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-secret-manager==2.24.0
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.73.0
idna==3.10
ImageHash==4.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonify==0.5
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
playwright==1.52.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyee==13.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.4
rsa==4.9.1
six==1.17.0
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
